# data/api-spec.json
{
    "components": {
        "schemas": {
...
            "ConsumptionLastTwoWeeks_Response": {
                "properties": {
                    "apartment_id": {
                        "title": "Apartment Id",
                        "type": "integer"
                    },
                    "consumption": {
                        "title": "Consumption",
                        "type": "number"
                    },
                    "date": {
                        "format": "date",
                        "title": "Date",
                        "type": "string"
                    },
                    "week_start": {
                        "format": "date",
                        "title": "Week Start",
                        "type": "string"
                    },
                    "weekday": {
                        "title": "Weekday",
                        "type": "string"
                    }
                },
                "required": [
                    "week_start",
                    "weekday",
                    "date",
                    "apartment_id",
                    "consumption"
                ],
                "title": "ConsumptionLastTwoWeeks_Response",
                "type": "object"
            },
...
            "DailyConsumption_Response": {
                "properties": {
                    "apartment_id": {
                        "title": "Apartment Id",
                        "type": "integer"
                    },
                    "consumption": {
                        "title": "Consumption",
                        "type": "number"
                    },
                    "date": {
                        "format": "date",
                        "title": "Date",
                        "type": "string"
                    }
                },
                "required": [
                    "consumption",
                    "apartment_id",
                    "date"
                ],
                "title": "DailyConsumption_Response",
                "type": "object"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "title": "Detail",
                        "type": "array"
                    }
                },
                "title": "HTTPValidationError",
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "type": "string"
                        },
                        "title": "Location",
                        "type": "array"
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                },
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError",
                "type": "object"
            }
        }
    },
    "info": {
        "description": "The OpenAPI schema for drop.energy",
        "title": "Drop API",
        "version": "0.1.0",
        "x-logo": {
            "url": "https://api.drop.energy/static/drop-logo.png"
        }
    },
    "openapi": "3.0.2",
    "paths": {
...
        "/kpi/consumption-daily/{apartment_id}/{date}": {
            "get": {
                "operationId": "get_consumption_kpi_consumption_daily__apartment_id___date__get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apartment_id",
                        "required": true,
                        "schema": {
                            "minimum": 0.0,
                            "title": "The ID of the apartment to get data for",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "path",
                        "name": "date",
                        "required": true,
                        "schema": {
                            "title": "The date to get data for",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DailyConsumption_Response"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Get Consumption",
                "tags": [
                    "consumption"
                ]
            }
        },
        "/kpi/consumption-weekly/{apartment_id}/{num_weeks}": {
            "get": {
                "operationId": "get_consumption_last_two_week_kpi_consumption_weekly__apartment_id___num_weeks__get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apartment_id",
                        "required": true,
                        "schema": {
                            "minimum": 0.0,
                            "title": "The ID of the apartment to get data for",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "path",
                        "name": "num_weeks",
                        "required": true,
                        "schema": {
                            "title": "The number of weeks to get data for",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ConsumptionLastTwoWeeks_Response"
                                    },
                                    "title": "Response Get Consumption Last Two Week Kpi Consumption Weekly  Apartment Id   Num Weeks  Get",
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Get Consumption Last Two Week",
                "tags": [
                    "consumption"
                ]
            }
        },
        "/kpi/hygiene-scores/{group_id}/{date}": {
            "get": {
                "operationId": "hygiene_scores_kpi_hygiene_scores__group_id___date__get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "group_id",
                        "required": true,
                        "schema": {
                            "title": "Group Id",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "path",
                        "name": "date",
                        "required": true,
                        "schema": {
                            "title": "Date",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Hygiene Scores",
                "tags": [
                    "kpi"
                ]
            }
        },
        "/kpi/sustainability-scores/{group_id}/{date}": {
            "get": {
                "operationId": "sustainability_scores_kpi_sustainability_scores__group_id___date__get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "group_id",
                        "required": true,
                        "schema": {
                            "title": "Group Id",
                            "type": "integer"
                        }
                    },
                    {
                        "in": "path",
                        "name": "date",
                        "required": true,
                        "schema": {
                            "title": "Date",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Sustainability Scores",
                "tags": [
                    "kpi"
                ]
            }
        },
        "/kpi/weekly-change/{apartment_id}": {
            "get": {
                "operationId": "get_user_weekly_trend_kpi_weekly_change__apartment_id__get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "apartment_id",
                        "required": true,
                        "schema": {
                            "minimum": 0.0,
                            "title": "The ID of the apartment to get data for",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get User Weekly Trend Kpi Weekly Change  Apartment Id  Get",
                                    "type": "number"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Get User Weekly Trend",
                "tags": [
                    "kpi"
                ]
            }
        },
...
    }
}